core/requirements.txt
Django==4.1.1
djangorestframework==3.13.1
django-debug-toolbar==3.6.0
django-phonenumber-field==7.0.0
phonenumbers==8.12.55
flake8==5.0.4
flake8-django==1.1.5
PyYAML==6.0
uritemplate==4.1.1
django-cors-headers==3.13.0
gunicorn==20.1.0
psycopg2-binary==2.9.5  # For dev postgres (for prod - psycopg2)